Added MovieDetail view
