+++ src/contexts/NavTitleContext.jsx
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+// Create the context
+export const NavTitleContext = createContext();
