+++ src/hooks/useAuth.jsx
@@ -1,5 +1,5 @@
-import { useContext } from "react";
 import { AuthContext } from "../contexts/AuthContext";
+import { useContext } from "react";
 
 export function useAuth() {
 	const context = useContext(AuthContext);
