+++ src/hooks/useMovies.jsx
@@ -1,0 +1,10 @@
+import { MoviesContext } from "../contexts/MoviesContext";
+import { useContext } from "react";
+
+export default function useMovies() {
+	const context = useContext(MoviesContext);
+	if (!context) {
+		throw new Error("useMovies must be used within a MoviesProvider");
+	}
+	return context;
+}
