+++ src/index.css
@@ -10,10 +10,10 @@ body {
 }
 
 nav {
-	@apply sticky top-0 bg-primary-800 rounded-md p-1;
+	@apply sticky top-0 mb-0 bg-primary-800 rounded-md;
 }
 nav ul {
-	@apply bg-primary-800 rounded-t-md p-1 flex flex-row justify-between;
+	@apply bg-primary-800 rounded-t-md p-1 flex sm:flex-row flex-col justify-between text-center;
 }
 
 nav .active {
@@ -26,15 +26,19 @@ div {
 
 /* Text */
 h1 {
-	@apply sticky ml-3 top-10 font-bold text-xl my-2 bg-primary-700;
+	@apply sticky mt-2 top-14 font-bold text-xl bg-primary-700;
 }
 
 h2 {
-	@apply font-bold text-lg my-2;
+	@apply font-bold text-lg mt-3;
 }
 
 h3 {
-	@apply font-bold my-2;
+	@apply font-bold py-0 my-0;
+}
+
+li {
+	@apply p-2;
 }
 
 button {
@@ -42,46 +46,49 @@ button {
 }
 
 /* Custom Classes - Please Sort Alphabetically */
-.ccForm-form-container{
-    @apply w-full rounded-lg border border-gray-200 bg-white p-4 shadow-sm dark:border-gray-700 dark:bg-gray-800 sm:p-6 lg:max-w-xl lg:p-8
-
+.ccForm-form-container {
+	@apply w-full rounded-lg border border-gray-200 bg-white p-4 shadow-sm dark:border-gray-700 dark:bg-gray-800 sm:p-6 lg:max-w-xl lg:p-8;
 }
 
-.ccForm-label{
-    @apply mb-2 block text-sm font-medium text-gray-900 dark:text-white
-
+.ccForm-label {
+	@apply mb-2 block text-sm font-medium text-gray-900 dark:text-white;
 }
 
-.ccForm-field{
-    @apply block w-full rounded-lg border border-gray-300 bg-gray-50 p-2.5 pe-10 text-sm text-gray-900 focus:border-primary-500 focus:ring-primary-500 dark:border-gray-600 dark:bg-gray-700 dark:text-white dark:placeholder:text-gray-400 dark:focus:border-primary-500 dark:focus:ring-primary-500
+.ccForm-field {
+	@apply block w-full rounded-lg border border-gray-300 bg-gray-50 p-2.5 pe-10 text-sm text-gray-900 focus:border-primary-500 focus:ring-primary-500 dark:border-gray-600 dark:bg-gray-700 dark:text-white dark:placeholder:text-gray-400 dark:focus:border-primary-500 dark:focus:ring-primary-500;
 }
 
-.infoform-div-container-centered{
-    display: flex;
-    justify-content: center;
-    align-items: center; 
-    flex-direction: column;
+.contents-container {
+	@apply min-h-full p-0;
 }
 
-.infoform-div-container{
-    @apply relative z-0 w-full mb-5 
+.infoform-div-container-centered {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	flex-direction: column;
 }
 
-.infoform-form-container{
-    @apply max-w-md mx-auto
+.infoform-div-container {
+	@apply relative z-0 w-full mb-5;
 }
 
-.infoform-input-field{
-    @apply block py-2.5 px-0 w-full text-sm text-gray-900 bg-transparent border-0 border-b-2 border-gray-300 appearance-none dark:text-white dark:border-gray-600 dark:focus:border-blue-500 focus:outline-none focus:ring-0 focus:border-blue-600
+.infoform-form-container {
+	@apply max-w-md mx-auto;
 }
 
-.infoform-input-label{
-    @apply peer-focus:font-medium absolute text-sm text-gray-500 dark:text-gray-400 duration-300 transform -translate-y-6 scale-75 top-3 -z-10 origin-[0] peer-focus:start-0 rtl:peer-focus:translate-x-1/4 peer-focus:text-blue-600 peer-focus:dark:text-blue-500 peer-placeholder-shown:scale-100 peer-placeholder-shown:translate-y-0 peer-focus:scale-75 peer-focus:-translate-y-6
-
+.infoform-input-field {
+	@apply block py-2.5 px-0 w-full text-sm text-gray-900 bg-transparent border-0 border-b-2 border-gray-300 appearance-none dark:text-white dark:border-gray-600 dark:focus:border-blue-500 focus:outline-none focus:ring-0 focus:border-blue-600;
 }
 
-.submit-button{
-
-    @apply text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm w-full sm:w-auto px-5 py-2.5 text-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800
+.infoform-input-label {
+	@apply peer-focus:font-medium absolute text-sm text-gray-500 dark:text-gray-400 duration-300 transform -translate-y-6 scale-75 top-3 -z-10 origin-[0] peer-focus:start-0 rtl:peer-focus:translate-x-1/4 peer-focus:text-blue-600 peer-focus:dark:text-blue-500 peer-placeholder-shown:scale-100 peer-placeholder-shown:translate-y-0 peer-focus:scale-75 peer-focus:-translate-y-6;
 }
 
+#root {
+	@apply min-h-full;
+}
+
+.submit-button {
+	@apply text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm w-full sm:w-auto px-5 py-2.5 text-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800;
+}
