+++ src/main.jsx
@@ -1,18 +1,21 @@
-import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import App from "./App.jsx";
+import App from "./App";
 import "./index.css";
 import { BrowserRouter } from "react-router-dom"; // Import BrowserRouter
 import AuthProvider from "./contexts/AuthProvider";
+import MoviesProvider from "./contexts/MoviesProvider";
+import NavTitleProvider from "./contexts/NavTitleProvider";
 
 const basename = import.meta.env.MODE === "development" ? "/" : "/frontend/";
 
 createRoot(document.getElementById("root")).render(
-	<StrictMode>
-		<AuthProvider>
-			<BrowserRouter basename={basename}>
-				<App />
-			</BrowserRouter>
-		</AuthProvider>
-	</StrictMode>
+	<AuthProvider>
+		<NavTitleProvider>
+			<MoviesProvider>
+				<BrowserRouter basename={basename}>
+					<App />
+				</BrowserRouter>
+			</MoviesProvider>
+		</NavTitleProvider>
+	</AuthProvider>
 );
